Refactor pen global variable to penPosition
Fix error where pen wraps around screen.

Currently the cursor is in the form of a global variable called pen
which is equivalent to the cell div the cursor is on. I wanted to find
a way to have a working cursor without the use of a global variable,
but found it difficult to do so.

My current approach to reduce the risk of using a global variable is to
make it just be a number, which represents the data-cell-number of the
cursor, as opposed to an entire cell div object as it was before.

This way, I can localize the pen object returned from document.query to
the scope of each function where it is needed, and be able to query the
proper cell div through the use of the penPosition.