Improve resetGrid time complexity
Although the current resetGrid works, I noticed that it takes a while
for the grid to actually reset.

I realized that the document.querySelector is a linear search for the
element, which results in O(n^2) runtime since I run an O(n) search for
all n elements of the grid.

By using document.querySelectorAll, which returns an nodeList of all
elements found in one pass or a single linear search, I thought I could
improve the runtime.

NodeList offers random access to elements, which is O(1), resulting in a
highest order of O(n) for the new resetGrid.

After making this change, the grid resets much faster.